requirements.txt
python-dotenv~=0.20.0
instagrapi~=1.16.20